refactor(test): take the acceptance-test helpers off the importable package
test_provider.go and test_provider_framework.go held only test helpers but were
compiled into the importable spacelift package, which leaked their SDKv2
helper/resource import into every importer. The no-drift migration tests need
terraform-plugin-testing/helper/resource, and both packages register a global
-sweep flag at init, so anything linking the two panics with "flag redefined:
sweep" before a single test runs. Renaming to _test.go keeps the helpers
available to package tests while taking that import off the public surface.

Also adds testStepsMux/testStepsMuxSequential, mirroring main.go's mux wiring, for
configs that reference a migrated Framework resource alongside one still on SDKv2.
The file keeps its //nolint:unused: which helpers have callers shifts as resources
migrate, so some are legitimately dormant at any given commit.
